requirements.txt
fastmcp>=0.1.0
fastapi>=0.128.4
httpx>=0.28.1
python-dotenv>=1.2.1
tomli>=2.4.0
tronpy>=0.6.2
uvicorn[standard]>=0.40.0